restructured routes to only contain routes
