+++ vidly-app/routes/customers.js
@@ -1,28 +1,7 @@
 const mongoose = require('mongoose');
 const express = require('express');
 const router = express.Router();
-const Joi = require('joi');
-
-const customerSchema = new mongoose.Schema({
-    isGold:{
-        type: Boolean,
-        required: true
-    },
-    name: {
-        type: String, 
-        required: true,
-        minlength: 1,
-        maxlength: 75
-    },
-    phone: {
-        type: String, 
-        required: true,
-        minlength: 1,
-        maxlength: 75
-    }
-});
-
-const Customer = mongoose.model('Customer', customerSchema)
+const {validate, Customer}= require('../models/customers');
 
 
 router.get('/', async (req, res) =>{
@@ -30,9 +9,8 @@ router.get('/', async (req, res) =>{
     res.send(customers)
 });
 
-
 router.post('/', async (req, res) => {
-    const { error } = validateCustomer(req.body);
+    const { error } = validate(req.body);
     if(error) return res.status(400).send(error.details[0].message);
 
     let customer = new Customer({
@@ -46,9 +24,8 @@ router.post('/', async (req, res) => {
     res.send(customer);
 });
 
-
 router.put('/:id', async (req, res) => {
-    const { error } = validateCustomer(req.body);
+    const { error } = validate(req.body);
     if(error) return res.status(400).send(error.details[0].message);
 
     const customer = await Customer.findByIdAndUpdate(req.params.id, 
@@ -63,8 +40,6 @@ router.put('/:id', async (req, res) => {
     if(!customer) return res.status(404).send('The Customer  you are looking for isn\'t here ');
 
     res.send(customer);
-
-
 });
 
 router.delete('/:id', async (req, res) => {
@@ -72,7 +47,7 @@ router.delete('/:id', async (req, res) => {
 
     if(!customer) return res.status(404).send('The Customer  you are looking for isn\'t here');
     res.send(customer);
-})
+});
 
 router.get('/:id', async (req, res) => {
     const customer = await Customer.findById(req.params.id);
@@ -81,14 +56,5 @@ router.get('/:id', async (req, res) => {
 
     res.send(customer);
 })
-
-function validateCustomer(customer){
-    const schema ={
-        isGold: Joi.boolean().required(),
-        name: Joi.string().min(1).max(75).required(),
-        phone: Joi.string().min(10).max(75).required(),
-    }
-    return Joi.validate(customer, schema);
-}
 
 module.exports = router;
