+++ vidly-app/routes/genres.js
@@ -1,21 +1,7 @@
 const mongoose = require('mongoose');
 const express = require('express');
 const router = express.Router();
-const Joi = require('joi');
-
-
-const Genre = mongoose.model('Genre', new mongoose.Schema({
-    name:{
-     type: String, 
-     required: true,
-     minlength: 5,
-     maxlength: 50
-     }
- })
-
-);
-
-
+const {Genre, validate} = require('../models/genres');
 
 // // show all genres
 router.get('/', async (req, res) =>{
@@ -23,11 +9,9 @@ router.get('/', async (req, res) =>{
     res.send(genres); 
 });
 
-
-
 // add a  genre to the genres array
 router.post('/', async (req, res) =>{
-    const { error } = validateGenre(req.body);
+    const { error } = validate(req.body);
     if(error) return res.status(400).send(error.details[0].message);
 
     let genre = new Genre({ name: req.body.name });
@@ -39,7 +23,7 @@ router.post('/', async (req, res) =>{
 
 // update genre
 router.put('/:id', async (req, res) =>{
-    const { error } = validateGenre(req.body);
+    const { error } = validate(req.body);
     if(error) return res.status(400).send(error.details[0].message);
 
     const genre = await Genre.findByIdAndUpdate(req.params.id, {name: req.body.name}, {
@@ -58,8 +42,6 @@ router.delete('/:id', async (req, res) =>{
     res.send(genre);
 });
 
-
-
 router.get('/:id', async (req, res) =>{
     const genre = await Genre.findById(req.params.id);
 
@@ -67,13 +49,5 @@ router.get('/:id', async (req, res) =>{
 
     res.send(genre);
 });
-
-
- function validateGenre(genre){
-    const schema ={
-        name: Joi.string().min(3).required()
-    }
-    return Joi.validate(genre, schema);
-}
 
 module.exports = router;
